githubapi: Detect "merged" state for comments on PRs.
Since the IssueCommentEvent itself doesn't contain enough information
for us to know whether the PR has been merged or not, we fetch
mentioned PRs (whether they've been merged). This is similar to us
fetching additional information about mentioned commits.

Then, we use the heuristic that if the PR is merged at present time,
and the event had "closed" state, then it's replaced by "merged" state
instead.

This has a small chance of false positives. Imagine a PR was closed,
then someone commented on it, then it was reopened and merged. So at
present time, the PR is merged. But when the comment was made, the PR
was in fact a closed PR, not a merged PR.

Fixing this false positive would require fetching the time at which the
PR was closed and comparing the event time to that. This should be done
when there's an accessible test case for it, because it's hard
otherwise. Also, other parts of the code might have the same flaw, and
should also be vetted at that time.

When we're using GraphQL API for fetching events, it'll likely be
viable and trivial to directly query whether the PR was merged at the
time of event.

Resolves the following TODO comment:

	// TODO: Detect "merged" state somehow? It's likely going to require making an API call.