githubapi: Perform style changes.
The purpose of this commit is to make the diff of next commit smaller.
It applies various style changes to try to make the code more readable
and consistent.

The two if err != nil statements after rc, err := s.fetchCommit are
merged into a single if/else if statement, to make it more clear that
the second if err != nil statement deals with the same error returned
from fetchCommit. Rename the *github.ErrorResponse variable to e to
avoid shadowing original error.