fs: Create eventsDir if it doesn't exist.
When writing the first log entry for a user, its eventsDir directory
may not exist yet. If writing fails, create the eventsDir directory
and try again.

This helps with the user experience of using this package for the
first time, by removing the need for the user to manually create
the eventsDir directory in advance.

Handle the not exist error in load in a roughly equivalent,
but more readable way.