githubapi: fetch module paths from go.mod files in GitHub repositories
If the latest commit of the default branch of a GitHub repository has
a go.mod file in the root directory, use the module path specified in it
instead of using the github.com/user/repo repository path. This is
a more accurate representation of the Go packages in said repository.

In the future, we can do more work to parse the issue/PR title prefix
and combine it with the module path to get the complete import path
of the relevant Go packages.