event, githubapi: set Event.Container to full import path
Parse the prefixed issue/change title to compute the full import path
of the affected package, and use that as the Event.Container value for
issue, change, and commit comment events.

For other event types, Event.Container continues to be the module path
corresponding to the repository root.

Documentat that the Create.Description field is optional.

Updates https://github.com/shurcooL/home/issues/26.